(no-ticket) Add Python compiler and runner
Add a Python compiler which "compiles" python files by generating
bytecode cache and packing it, together with the source files,
into an archive.

Add a Python3 file runner, which unpacks python code from archives
and runs it.

Pass /dev/zero as /dev/urandom in the sandbox both when compiling
and running.

To generate bytecode cache, use python -m compileall

The performance improvement is negligible most of the time,
but this prevents Python from trying to create those files
at runtime, which triggers RV due to open(..., O_WRONLY, ...)

This also allows detecting syntax error at compile time,
and producing CE with a user-visible error message, instead of a RE.

Use of undefined variables, and so on, will still lead to RE,
and there's nothing we can do about it.

Packing the whole directory into a single archive makes it easy
to deliver the bytecode cache together with the python source,
and will make it easier to add support for tasks with libraries
in the future.

We use tar archives, because they preserve timestamps,
which prevents Python from trying to re-generate the pyc files.

Python 3 needs /dev/urandom, but instead of bind-mounting real
/dev/urandom into the sandbox, we mount /dev/zero as /dev/urandom,
so that the "randomness" is fixed.

This should make python runs more deterministic.

Change-Id: I22d7ca6b141d35f75c15a1a1fe64772d5c08e0ac