(no-ticket) Make proot's bind-mounting compatible with sio2jail sandboxes
Sio2jail sandboxes contain empty files or directories
at mountpoints, because of the way bind-mounts work.

This means that if /bin/sh is to be bind-mounted into them,
they will contain an empty /bin/sh,
so checking if /bin/sh exists is not enough to tell
if it's working, or if just an empty file
waiting for a bind-mount.

In PRootSandboxExecutor check whether /bin/sh is executable
and bind-mount it if it's not.

Change-Id: I0d491ec35d0c2011250550ead18728516e99f21e