(no-ticket) Make null_ctx_manager reentrant
A generator-based context manager - as implemented by
contextlib.contextmanager - is not reentrant and not reusable,
i.e. each of __enter__ and __exit__ can be called only once.

For sandboxes, we need reentrant and reusable context managers,
because we often use the same sandbox object multiple times,
or, from within a with statement, call a function which has a with
statement on the same sandbox object.

Make null_ctx_manager implement context manager on its own,
instead of using contextlib, so that it can be used as a dummy sandbox.

Change-Id: If5fa0a337c6879d3d4f1b4d19af8f8c2931f47f0