(no-ticket) Add swig support to python compiler
Assume each .i file in extra_execution_files is meant to be used
with swig to generate python bindings for a .cpp file with the
same name.

Use swig and gcc from the python sandbox to generate the bindings
and build a python extension from each such pair of files,
and make the resulting .so and .py files available
to the contestant's program.

Change-Id: Ia4b1b15727bd9e964536f9f45c6ad50d446e939f